Add comments and reorg test location
