+++ crates/sparrow-instructions/src/evaluators/map.rs
@@ -1,2 +1,1 @@
-// pub(super) mod get_large_string_to_primitive;
 pub(super) mod get_string_to_primitive;
