+++ crates/sparrow-main/tests/e2e/basic_error_tests.rs
@@ -1088,32 +1088,3 @@ async fn test_unsupported_output_type_with_nested_formulas_2() {
           - ""
     "###);
 }
-
-#[tokio::test]
-async fn test_swapped_args_for_get_map() {
-    insta::assert_yaml_snapshot!(QueryFixture::new("{ f1: get(Input.e0, \"f1\") }")
-        .run_to_csv(&collection_data_fixture().await).await.unwrap_err(), @r###"
-    ---
-    code: Client specified an invalid argument
-    message: 1 errors in Fenl statements; see diagnostics
-    fenl_diagnostics:
-      - severity: error
-        code: E0010
-        message: Invalid argument type(s)
-        formatted:
-          - "error[E0010]: Invalid argument type(s)"
-          - "  --> Query:1:7"
-          - "  |"
-          - "1 | { f1: get(Input.e0, \"f1\") }"
-          - "  |       ^^^           ---- Actual type: string"
-          - "  |       |              "
-          - "  |       Invalid types for parameter 'map' in call to 'get'"
-          - "  |"
-          - "  --> built-in signature 'get<K: key, V: any>(key: K, map: map<K, V>) -> V':1:34"
-          - "  |"
-          - "1 | get<K: key, V: any>(key: K, map: map<K, V>) -> V"
-          - "  |                                  --------- Expected type: map<K, V>"
-          - ""
-          - ""
-    "###);
-}
