+++ crates/sparrow-main/tests/e2e/map_tests.rs
@@ -1,0 +1,7 @@
+//! Basic e2e tests for the map functions.
+
+use indoc::indoc;
+use sparrow_api::kaskada::v1alpha::TableConfig;
+use uuid::Uuid;
+
+use crate::{DataFixture, QueryFixture};
